+++ src/sonic_ax_impl/mibs/__init__.py
@@ -4,6 +4,8 @@ import re
 from swsssdk import SonicV2Connector
 from swsssdk import port_util
 from swsssdk.port_util import get_index, get_index_from_str
+from ax_interface.mib import MIBUpdater
+from ax_interface.util import oid2tuple
 from sonic_ax_impl import logger
 
 COUNTERS_PORT_NAME_MAP = b'COUNTERS_PORT_NAME_MAP'
@@ -364,3 +366,54 @@ def get_transceiver_sensor_sub_id(ifindex, sensor):
 
     transceiver_oid, = get_transceiver_sub_id(ifindex)
     return (transceiver_oid + SENSOR_PART_ID_MAP[sensor], )
+
+class RedisOidTreeUpdater(MIBUpdater):
+    def __init__(self, prefix_str):
+        super().__init__()
+
+        self.db_conn = init_db()
+        if prefix_str.startswith('.'):
+            prefix_str = prefix_str[1:]
+        self.prefix_str = prefix_str
+
+    def get_next(self, sub_id):
+        """
+        :param sub_id: The 1-based sub-identifier query.
+        :return: the next sub id.
+        """
+        raise NotImplementedError
+
+    def reinit_data(self):
+        """
+        Subclass update loopback information
+        """
+        pass
+
+    def update_data(self):
+        """
+        Update redis (caches config)
+        Pulls the table references for each interface.
+        """
+        self.oid_list = []
+        self.oid_map = {}
+
+        self.db_conn.connect(SNMP_OVERLAY_DB)
+        keys = self.db_conn.keys(SNMP_OVERLAY_DB, self.prefix_str + '*')
+        # TODO: fix db_conn.keys to return empty list instead of None if there is no match
+        if keys is None:
+            keys = []
+
+        for key in keys:
+            key = key.decode()
+            oid = oid2tuple(key, dot_prefix=False)
+            self.oid_list.append(oid)
+            value = self.db_conn.get_all(SNMP_OVERLAY_DB, key)
+            if value[b'type'] in [b'COUNTER_32']:
+                self.oid_map[oid] = int(value[b'data'])
+
+        self.oid_list.sort()
+
+    def get_oidvalue(self, oid):
+        if oid not in self.oid_map:
+            return None
+        return self.oid_map[oid]
